BUG: Fix PointSetToPointSetMetric Thread Safety
This reverts to the multitrheaded PointSetToPointSetMetric and fixes a
thread safety issue when using itkComositeTransform.

the method GetNumberOfLocalParameters in itkCompositeTransform is not
threadsafe due to the timestamp being updated before the variable, which
can results in returning a wrong number to a concurrent accessing thread.

This fix moves GetNumberOfLocalParameters calls from threads.

Additionally it fixes a wrong size of derivative for one of the possible
return paths.